added verify tx pages
