+++ website/src/less/main.less
@@ -335,6 +335,10 @@ footer {
 		background-color: @light-grey-color;
 	}
 
+	div.text-result {
+		word-wrap: break-word;
+	}
+
 	@media (max-width: 768px) {
 		text-align: center;
