+++ website/src/views/index.pug
@@ -33,11 +33,11 @@ html
 					div.col-lg-3.col-md-4.col-sm-5
 						div.smaller
 							span Are you a developer?
-						a.btn-rounded(onClick="") Check status
+						a.btn-rounded(href="./developers.html") Check status
 					div.col-lg-3.col-md-4.col-sm-5
 						div.smaller
 							span Everybody can be a relayer#[br.hidden-xs] and earn ETH
-						button.btn-rounded.btn-red(onClick="") Become a relayer
+						a.btn-rounded.btn-red(href="./become-a-relayer.html") Become a relayer
 		section
 			div.container.text-center
 				h2 How it works?
