+++ website/src/views/partials/status.pug
@@ -1,4 +1,4 @@
-div#statusPage.hidden
+div#statusPage
 	hr
 	div.row
 		div.col-lg-3
